requirements.txt
langchain
langgraph
google-generativeai
langchain_community
python-dotenv
pdfplumber
langchain-groq
streamlit
ipython
fpdf
-e .